Reorganized unit tests to follow the OpenStack layout
