+++ st2mistral/tests/unit/functions/test_time.py
@@ -15,11 +15,13 @@
 
 from st2mistral.tests.unit import test_function_base as base
 
+from yaql.language import factory
+YAQL_ENGINE = factory.YaqlFactory().create()
 
-class JinjaUtilsTimeFunctionTestCase(base.JinjaFunctionTestCase):
+
+class JinjaTimeTestCase(base.JinjaFunctionTestCase):
 
     def test_to_human_time_function(self):
-
         template = '{{ to_human_time_from_seconds(_.k1) }}'
         result = self.eval_expression(template, {'k1': 12345})
         self.assertEqual(result, '3h25m45s')
@@ -29,3 +31,24 @@ class JinjaUtilsTimeFunctionTestCase(base.JinjaFunctionTestCase):
 
         with self.assertRaises(AssertionError):
             self.eval_expression(template, {'k1': 'stuff'})
+
+
+class YAQLTimeTestCase(base.YaqlFunctionTestCase):
+
+    def test_to_human_time_from_seconds(self):
+        expression = 'to_human_time_from_seconds($.input_int)'
+        result = YAQL_ENGINE(expression).evaluate(
+            context=self.get_yaql_context({"input_int": 12345})
+        )
+        self.assertEqual(result, '3h25m45s')
+
+        expression = 'to_human_time_from_seconds($.input_int)'
+        result = YAQL_ENGINE(expression).evaluate(
+            context=self.get_yaql_context({"input_int": 0})
+        )
+        self.assertEqual(result, '0s')
+
+        with self.assertRaises(AssertionError):
+            YAQL_ENGINE('to_human_time_from_seconds($.input_int)').evaluate(
+                context=self.get_yaql_context({"input_int": "foo"})
+            )
